requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud==0.34.0
langchain==0.1.16
langchain-openai==0.1.1
numpy==1.26.4
oauth2client==4.1.3
pypdf==4.0.1
requests==2.31.0
vertexai==1.49.0
cryptography==3.1